requirements.txt
alembic==1.8.1
aniso8601==9.0.1
apiclient==1.0.4
APScheduler==3.9.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-APScheduler==1.12.4
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
httplib2==0.20.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
oauth2client==4.1.3
packaging==21.3
protobuf==4.21.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
tzdata==2022.4
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
